pyproject.toml
[project]
name = "urfu-robot"
version = "0.1.0"
description = ""
authors = [
    {name = "liinaD3210",email = "vodenikov.daniil@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "streamlit (>=1.45.1,<2.0.0)",
    "langchain-google-genai (>=2.1.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "requests (>=2.32.3,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "pypdf (>=5.5.0,<6.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "urfu_robot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
